remove fields and options of removed forms (migration)

+++ back/db/migrate/20220830144847_move_custom_forms_to_participation_context.rb
@@ -17,14 +17,31 @@ class MoveCustomFormsToParticipationContext < ActiveRecord::Migration[6.1]
       WHERE participation_context_id IS NULL
     SQL
     delete_custom_form_ids = delete_custom_form_results.pluck 'id'
-    ActiveRecord::Base.connection.execute <<~SQL.squish
-      DELETE FROM custom_forms
-      WHERE id IN (#{delete_custom_form_ids.join(', ')})
-    SQL
-    ActiveRecord::Base.connection.execute <<~SQL.squish
-      DELETE FROM custom_fields
-      WHERE resource_id IN (#{delete_custom_form_ids.join(', ')})
-    SQL
+    if delete_custom_form_ids.present?
+      sql_form_ids = delete_custom_form_ids.map { |id| "'#{id}'" }.join(', ')
+      ActiveRecord::Base.connection.execute <<~SQL.squish
+        DELETE FROM custom_forms
+        WHERE id IN (#{sql_form_ids})
+      SQL
+
+      delete_custom_field_results = ActiveRecord::Base.connection.execute <<~SQL.squish
+        SELECT id
+        FROM custom_fields
+        WHERE resource_id IN (#{sql_form_ids})
+      SQL
+      delete_custom_field_ids = delete_custom_field_results.pluck 'id'
+      if delete_custom_field_ids.present?
+        sql_field_ids = delete_custom_field_ids.map { |id| "'#{id}'" }.join(', ')
+        ActiveRecord::Base.connection.execute <<~SQL.squish
+          DELETE FROM custom_field_options
+          WHERE custom_field_id IN (#{sql_field_ids})
+        SQL
+        ActiveRecord::Base.connection.execute <<~SQL.squish
+          DELETE FROM custom_fields
+          WHERE id IN (#{sql_field_ids})
+        SQL
+      end
+    end
     change_column_null :custom_forms, :participation_context_id, false
 
     add_column :custom_forms, :participation_context_type, :string
